Falta un caso en back, el reporte de ventas

+++ facturaPOS/src/main/java/com/example/facturaPOS/repository/CategoriaRepository.java
@@ -1,9 +1,10 @@
 package com.example.facturaPOS.repository;
 
-import com.example.facturaPOS.model.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.example.facturaPOS.model.Categoria;
+
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
 }
