+++ facturaPOS/src/main/java/com/example/facturaPOS/repository/MesaRepository.java
@@ -1,9 +1,10 @@
 package com.example.facturaPOS.repository;
 
-import com.example.facturaPOS.model.Mesa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.example.facturaPOS.model.Mesa;
+
 @Repository
-public interface MesaRepository extends JpaRepository<Mesa, Integer> {
+public interface MesaRepository extends JpaRepository<Mesa, Long> {
 }
