+++ facturaPOS/src/main/java/com/example/facturaPOS/repository/PagoRepository.java
@@ -1,9 +1,10 @@
 package com.example.facturaPOS.repository;
 
-import com.example.facturaPOS.model.Pago;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.example.facturaPOS.model.Pago;
+
 @Repository
-public interface PagoRepository extends JpaRepository<Pago, Integer> {
+public interface PagoRepository extends JpaRepository<Pago, Long> {
 }
